+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.15 branch
 ### 2.15.0
+* removed `IntegrationTestCase::loadAllFixtures()` method;
 * removed `am()` global function.
 
 ## 2.14 branch